Remove k command and use dnx instead
