+++ DNX.sln
@@ -1,13 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.22712.0
+VisualStudioVersion = 14.0.22717.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "scripts", "scripts", "{189961D1-DFF4-406A-9D42-39B61221A73A}"
 	ProjectSection(SolutionItems) = preProject
-		scripts\dnx.cmd = scripts\dnx.cmd
 		scripts\k-crossgen.cmd = scripts\k-crossgen.cmd
-		scripts\k.cmd = scripts\k.cmd
 		makefile.shade = makefile.shade
 	EndProjectSection
 EndProject
@@ -110,7 +108,7 @@ Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.Runtime
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "NuGet.Packaging.Core.Types", "submodules\NuGet.Packaging\src\NuGet.Packaging.Core.Types\NuGet.Packaging.Core.Types.xproj", "{46275450-3B62-4E89-8F7E-3FE2EE8513D2}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "NuGet.ProjectModel", "submodules\NuGet.Packaging\src\NuGet.ProjectModel\NuGet.ProjectModel.xproj", "{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}"
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "NuGet.ProjectModel", "submodules\NuGet.Packaging\src\NuGet.ProjectModel\NuGet.ProjectModel.xproj", "{18006692-AFCC-49FD-95F6-66E8A0C49E9B}"
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "NuGet.LibraryModel", "submodules\NuGet.Packaging\src\NuGet.LibraryModel\NuGet.LibraryModel.xproj", "{13883E8E-7DE1-4EDD-8E4A-C5357BA8CD81}"
 EndProject
@@ -804,26 +802,26 @@ Global
 		{46275450-3B62-4E89-8F7E-3FE2EE8513D2}.Release|x64.Build.0 = Release|Any CPU
 		{46275450-3B62-4E89-8F7E-3FE2EE8513D2}.Release|x86.ActiveCfg = Release|Any CPU
 		{46275450-3B62-4E89-8F7E-3FE2EE8513D2}.Release|x86.Build.0 = Release|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Debug|Win32.ActiveCfg = Debug|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Debug|Win32.Build.0 = Debug|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Debug|x64.Build.0 = Debug|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Debug|x86.Build.0 = Debug|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Release|Win32.ActiveCfg = Release|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Release|Win32.Build.0 = Release|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Release|x64.ActiveCfg = Release|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Release|x64.Build.0 = Release|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Release|x86.ActiveCfg = Release|Any CPU
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5}.Release|x86.Build.0 = Release|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Debug|Win32.ActiveCfg = Debug|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Debug|Win32.Build.0 = Debug|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Debug|x64.Build.0 = Debug|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Debug|x86.Build.0 = Debug|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Release|Win32.ActiveCfg = Release|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Release|Win32.Build.0 = Release|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Release|x64.ActiveCfg = Release|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Release|x64.Build.0 = Release|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Release|x86.ActiveCfg = Release|Any CPU
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B}.Release|x86.Build.0 = Release|Any CPU
 		{13883E8E-7DE1-4EDD-8E4A-C5357BA8CD81}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{13883E8E-7DE1-4EDD-8E4A-C5357BA8CD81}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{13883E8E-7DE1-4EDD-8E4A-C5357BA8CD81}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
@@ -1052,7 +1050,7 @@ Global
 		{AF8C0E0F-D6F2-4F22-BF0D-0E904F46DD95} = {A6FE6180-B68C-447F-9CD6-66287B329B80}
 		{209FE200-D4E8-4E52-9AEA-7F77A10696A8} = {AF391791-F4B7-41AC-8F08-9485DAC543C5}
 		{46275450-3B62-4E89-8F7E-3FE2EE8513D2} = {CAF3AE3C-D013-43E7-9AFD-AC6956B2999F}
-		{F013E43F-B6D5-4F59-ACF0-EECEC2C794F5} = {CAF3AE3C-D013-43E7-9AFD-AC6956B2999F}
+		{18006692-AFCC-49FD-95F6-66E8A0C49E9B} = {CAF3AE3C-D013-43E7-9AFD-AC6956B2999F}
 		{13883E8E-7DE1-4EDD-8E4A-C5357BA8CD81} = {CAF3AE3C-D013-43E7-9AFD-AC6956B2999F}
 		{CF8E1753-EE98-410F-BB4B-6624B19C6B64} = {CAF3AE3C-D013-43E7-9AFD-AC6956B2999F}
 		{24DE9E2C-4ADC-48E6-ADF0-1E2B00596ECD} = {CAF3AE3C-D013-43E7-9AFD-AC6956B2999F}
