+++ test/Microsoft.Framework.ApplicationHost.FunctionalTests/project.json
@@ -1,5 +1,6 @@
 {
     "dependencies": {
+        "Bootstrapper.FunctionalTests": "1.0.0-*",
         "Microsoft.Framework.FunctionalTestUtils": "1.0.0-*",
         "Microsoft.Framework.Runtime.Common": "1.0.0-*",
         "xunit.runner.aspnet": "2.0.0-aspnet-*"
